Move `BloodPressureModule` to `di` package
